vod: use time instead of position for seek requests
RTSP uses NPT in seconds, so don't convert for nothing...

+++ modules/stream_out/rtp.h
@@ -93,7 +93,7 @@ void CloseVoD( vlc_object_t * );
 void vod_start(vod_media_t *p_media, const char *psz_session);
 void vod_toggle_pause(vod_media_t *p_media, const char *psz_session);
 void vod_stop(vod_media_t *p_media, const char *psz_session);
-void vod_seek(vod_media_t *p_media, const char *psz_session, float time);
+void vod_seek(vod_media_t *p_media, const char *psz_session, int64_t time);
 
 const char *vod_get_mux(const vod_media_t *p_media);
 int vod_init_id(vod_media_t *p_media, const char *psz_session, int es_id,
